+++ app/src/main/java/com/gojol/notto/ui/todo/dialog/TodoAlarmPeriodDialog.kt
@@ -1,23 +1,22 @@
-package com.gojol.notto.ui.home.dialog
+package com.gojol.notto.ui.todo.dialog
 
 import android.content.Context
 import android.view.LayoutInflater
 import androidx.databinding.DataBindingUtil
 import com.gojol.notto.R
-import com.gojol.notto.databinding.DialogTodoRepeatTypeBinding
+import com.gojol.notto.databinding.DialogTodoAlarmPeriodBinding
 
 
-class TodoRepeatTypeDialog(context: Context) : TodoBaseDialogImpl(context) {
-    private val binding: DialogTodoRepeatTypeBinding =
+class TodoAlarmPeriodDialog(context: Context) : TodoBaseDialogImpl(context) {
+    private val binding: DialogTodoAlarmPeriodBinding =
         DataBindingUtil.inflate(
-            LayoutInflater.from(context), R.layout.dialog_todo_repeat_type,
+            LayoutInflater.from(context), R.layout.dialog_todo_alarm_period,
             null,
             false
         )
 
     init {
         setBinding(binding)
-        setDialog()
         initClickListener()
     }
