+++ app/src/main/java/com/gojol/notto/ui/todo/dialog/TodoRepeatTypeDialog.kt
@@ -1,22 +1,23 @@
-package com.gojol.notto.ui.home.dialog
+package com.gojol.notto.ui.todo.dialog
 
 import android.content.Context
 import android.view.LayoutInflater
 import androidx.databinding.DataBindingUtil
 import com.gojol.notto.R
-import com.gojol.notto.databinding.DialogTodoAlarmPeriodBinding
+import com.gojol.notto.databinding.DialogTodoRepeatTypeBinding
 
 
-class TodoAlarmPeriodDialog(context: Context) : TodoBaseDialogImpl(context) {
-    private val binding: DialogTodoAlarmPeriodBinding =
+class TodoRepeatTypeDialog(context: Context) : TodoBaseDialogImpl(context) {
+    private val binding: DialogTodoRepeatTypeBinding =
         DataBindingUtil.inflate(
-            LayoutInflater.from(context), R.layout.dialog_todo_alarm_period,
+            LayoutInflater.from(context), R.layout.dialog_todo_repeat_type,
             null,
             false
         )
 
     init {
         setBinding(binding)
+        setDialog()
         initClickListener()
     }
