+++ app/src/main/res/layout/dialog_todo_alarm_period.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="dialog"
-            type="com.gojol.notto.ui.home.dialog.TodoAlarmPeriodDialog" />
+            type="com.gojol.notto.ui.todo.dialog.TodoAlarmPeriodDialog" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
