+++ app/src/main/res/layout/dialog_todo_repeat_time.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="dialog"
-            type="com.gojol.notto.ui.home.dialog.TodoRepeatTimeDialog" />
+            type="com.gojol.notto.ui.todo.dialog.TodoRepeatTimeDialog" />
     </data>
 
     <androidx.core.widget.NestedScrollView
