+++ app/src/main/res/layout/dialog_todo_repeat_type.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="dialog"
-            type="com.gojol.notto.ui.home.dialog.TodoRepeatTypeDialog" />
+            type="com.gojol.notto.ui.todo.dialog.TodoRepeatTypeDialog" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
