+++ app/src/main/res/layout/item_dialog_bottom_button.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="dialog"
-            type="com.gojol.notto.ui.home.dialog.TodoBaseDialogImpl" />
+            type="com.gojol.notto.ui.todo.dialog.TodoBaseDialogImpl" />
 
     </data>
